feat(#93): refactor: supervisor/ — fix anti-patterns across modules
