+++ .pi/extensions/supervisor/extensions.ts
@@ -47,11 +47,12 @@ export function resolveExtensions(extensionsRaw: string | undefined): string[] {
 
 let _extToolsCache: Map<string, string[]> | null = null;
 
-export function discoverExtensionTools(): Map<string, string[]> {
+export function discoverExtensionTools(cwd?: string): Map<string, string[]> {
 	if (_extToolsCache) return _extToolsCache;
 
 	const map = new Map<string, string[]>();
-	const extDir = resolvePath(process.cwd(), ".pi/extensions");
+	const baseCwd = cwd || process.cwd();
+	const extDir = resolvePath(baseCwd, ".pi/extensions");
 
 	let files: string[];
 	try {
@@ -94,7 +95,11 @@ export function discoverExtensionTools(): Map<string, string[]> {
  * Merge agent-declared tools with tools from agent's extensions.
  * Returns a comma-separated string for --tools flag.
  */
-export function resolveTools(agentTools: string, extNamesRaw: string | undefined): string {
+export function resolveTools(
+	agentTools: string,
+	extNamesRaw: string | undefined,
+	cwd?: string,
+): string {
 	const toolSet = new Set(
 		agentTools
 			.split(",")
@@ -103,7 +108,7 @@ export function resolveTools(agentTools: string, extNamesRaw: string | undefined
 	);
 
 	if (extNamesRaw && extNamesRaw.trim()) {
-		const extToolsMap = discoverExtensionTools();
+		const extToolsMap = discoverExtensionTools(cwd);
 		const extNames = extNamesRaw
 			.split(",")
 			.map((s) => s.trim())
